Modernize strategy test suite and implement ADS v3.0 compatibility shims
- Resolved 18 Pandas warnings (fillna, dtypes) across core engine and strategies.
- Implemented legacy compatibility shims in EnhancedBaseStrategy and StrategyCorrelationAnalyzer to support ADS v2.0 APIs on v3.0 engine.
- Unskipped and fixed 40 strategy unit tests, including infrastructure and legacy position management tests.
- Updated StrategyManager pipeline tests with correct EnrichedMarketData mocks.
- Converted strategy position management to async to align with orchestrator design.
- Verified 100% pass rate (552/552) for strategy unit tests.

+++ tests/unit/strategies/test_strategy_advanced_coverage.py
@@ -443,7 +443,7 @@ class TestBreakoutAdvanced:
         current_price = recent_high * 1.02  # Above resistance
         df.loc[df.index[-1], 'close'] = current_price
         df.loc[df.index[-1], 'high'] = current_price
-        df.loc[df.index[-1], 'volume'] = df['volume'].mean() * 2.0  # High volume
+        df.loc[df.index[-1], 'volume'] = int(df['volume'].mean() * 2.0)  # High volume
 
         strategy.market_data = enriched_data
 
@@ -466,7 +466,7 @@ class TestBreakoutAdvanced:
         current_price = recent_low * 0.98  # Below support
         df.loc[df.index[-1], 'close'] = current_price
         df.loc[df.index[-1], 'low'] = current_price
-        df.loc[df.index[-1], 'volume'] = df['volume'].mean() * 2.0  # High volume
+        df.loc[df.index[-1], 'volume'] = int(df['volume'].mean() * 2.0)  # High volume
 
         strategy.market_data = enriched_data
 
@@ -501,7 +501,7 @@ class TestBreakoutAdvanced:
         recent_high = df['high'].tail(20).max()
         current_price = recent_high * 1.02  # Above resistance
         df.loc[df.index[-1], 'close'] = current_price
-        df.loc[df.index[-1], 'volume'] = df['volume'].mean() * 0.5  # Low volume
+        df.loc[df.index[-1], 'volume'] = int(df['volume'].mean() * 0.5)  # Low volume
 
         strategy.market_data = enriched_data
